* Allow int to be passed as value

+++ Wazzy/Types/WASMType.cs
@@ -13,7 +13,7 @@ namespace Wazzy.Types
             _supportedTypes = new[] { typeof(int), typeof(long), typeof(float), typeof(double) };
         }
 
-        public static Type GetType(byte valueTypeId) => _supportedTypes[127 - valueTypeId];
+        public static Type GetType(int valueTypeId) => _supportedTypes[127 - valueTypeId];
         public static bool IsSupportedType(int valueTypeId) => valueTypeId >= 124 && valueTypeId <= 127;
     }
 }
